requirements.txt
fastapi[all]
uvicorn
Pillow
opencv-python
numpy
